feat(Backend-wallet): add wallet & image generation APIs, refactor services and DB schema
	•	Introduced WalletController and GenerateController with RESTful endpoints for balance, transactions, and generation tasks
	•	Added DTOs (BalanceDto, TransactionDto, GenerateRequestDto, etc.) and service interfaces (IWalletService, IGenerateService)
	•	Updated IgDbContext to include Transactions table and model configuration
	•	Refactored provider registration to singleton; improved service scoping
	•	Removed image generation from ConversationController and updated documentation
	•	Changed Credits in ProfileDto from int → decimal

+++ ImageGenerator/Models/User.cs
@@ -38,7 +38,7 @@ public class User: ModelBase
     /// <summary>
     /// The number of credits the user has for image generation.
     /// </summary>
-    public int Credits { get; set; } = 0;
+    public decimal Credits { get; set; } = 0;
 
     /// <summary>
     /// The timestamp when the user last claimed their credits.
